# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zenkatu Streaming API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /usuarios/{uid}:
    get:
      operationId: getUsuario
      tags: [usuarios]
      summary: Get user profile
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "404":
          description: Not found
    patch:
      operationId: updateUsuario
      tags: [usuarios]
      summary: Update user profile
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /usuarios:
    post:
      operationId: upsertUsuario
      tags: [usuarios]
      summary: Create or update user on login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
      responses:
        "200":
          description: User upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /usuarios/{uid}/avatar:
    post:
      operationId: uploadAvatar
      tags: [usuarios]
      summary: Upload user avatar (base64 WebP)
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarUpload"
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
...
components:
  schemas:
...
    Usuario:
      type: object
      required:
        - id
        - uid
        - email
      properties:
        id:
          type: integer
        uid:
          type: string
        email:
          type: string
        username:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    UsuarioInput:
      type: object
      required:
        - uid
        - email
      properties:
        uid:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
    UsuarioUpdate:
      type: object
      properties:
        username:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
    AvatarUpload:
      type: object
      required:
        - imageData
      properties:
        imageData:
          type: string
          description: Base64-encoded WebP image data
...