# lib/api-zod/src/generated/types/avatarUpload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zenkatu Streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface AvatarUpload {
  /** Base64-encoded WebP image data */
  imageData: string;
}