# lib/api-zod/src/generated/types/usuarioInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zenkatu Streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface UsuarioInput {
  uid: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}